git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1005 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/Array.h
@@ -424,7 +424,7 @@ public:
      */
    void trimToSize() {
        if (size() != capacity()) {
-           int oldNum = numAllocated;
+           size_t oldNum = numAllocated;
            numAllocated = size();
            realloc(oldNum);
        }
@@ -740,8 +740,8 @@ public:
        sequence, a value at least as large as size()" 
        For compatibility with std::vector.
    */
-   size_t capacity() const {
-       return numAllocated;
+    int capacity() const {
+        return (int)numAllocated;
    }
 
    /**
